ramiro branch: Refs. issue 35. Implement a 'ss_loaddata' management command.
This is done by means of a 'sql_server.extra' app that should be added to
INSTALLED APPS. ss_loaddata is similar to the Django built-in loaddata
but allows loading of fixtures with forward references to a MS SQL
Server DB.

This  has been implemented by detecting the tables affected by a fixture
and activating-deactivating constraint checks on a per-table basis
using 'ALTER <table> [WITH CHECK] [NO]CHECK CONSTRAINT ALL'.

Also, transparent replacement of the 'loaddata' command with 'ss_loaddata' when
the sql_server.extra app is instaled and the DB backend is SQL Server has been
implemented.

The interface of the implementation is something that could change in the future
depending on the outcome of Django ticket 3615 (possibly making innecesary the
use of a custom management command).